requirements.txt
pillow
streamlit
streamlit_lottie
streamlit_option_menu
